Change hook methods so that if a tag/method is specified it overwrites an existing hook block with the same tag/method
Assume:

  class Artist < Sequel::Model
    def do_stuff
      ...
    end
    before_save :do_stuff
    before_save :do_stuff
  end

Before:

  Artist.new.save # do_stuff called twice!

After:

  Artist.new.save # do_stuff called once

It's unlikely you would do the above in real code, but if you have
an application that does automatic reloading of source (such as
Ramaze), this will prevent it adding to the hooks everytime it
reloads the source file.

Note that you can add tags with blocks, to prevent the reloading:

  class Artist < Sequel::Model
    before_save(:do_stuff){...}
  end

If you use both tags and methods with the same name, the last
one wins.

tags/methods are stored per model, so this works the way you
expect:

  class Artist < Sequel::Model
    before_save(:do_stuff){...}
  end
  class BigArtist < Artist
    before_save(:do_stuff){...}
  end
  BigArtist.new.save # calls both the Artist and BigArtist blocks