Allow an array of arrays instead of a hash when specifying conditions
This was previously not allowed, as the code assumed the array was in the
[string, param1, param2, ...] format.  The code now handles an array the same
as before if the first element is a string, and treats it like a hash if not.

The need for this is to be able to specify multiple conditions on the same
column in the joined table when doing a join.  Previously, this was impossible,
as the code needed a hash and specifying multiple conditions would use the same
hash key, so one of the conditions would be ignored.

A side benefit to this is that you can have more predictable specs because
the array of arrays will always give the conditions in the same order in the
resulting SQL.