Refactor and beef up Dataset#first and Dataset#last, with some change in functionality
Make first(1) return an array of zero or one records instead of a single
record, since it will mostly likely be used as first(i).

If it block is given, it is used to filter records, but it no longer ignores
arguments.  If you pass a block and an integer argument, you get an array of
records filtered by the block and limited to the integer given. If you pass
a block and a non-integer argument(s), both the block and the arguments are
used to filter the records.

first with no arguments still returns a single record.

Last now just calls reverse and then first with the block and arguments it
was given, so it's much more powerful now.

Make single_record always limit the dataset to 1 record (why fetch records if
you are going to ignore them?)

Slight refactoring of dataset/convenience.rb (got sidetracked with first/last
while I was refactoring)