Add limited column reflection support, tested on PostgreSQL, MySQL, and SQLite
This supports finds the name, type, default value, NULL allowability, maximum
character length, and numeric precision of the columns for a given table. You
use it like:

  DB.schema_for_table(:artists)

The object returned is an associated array (array with two fields, one a key
and the other a value) instead of a hash to preserve column ordering.

This support isn't used by anything yet, and needs wider testing to catch more
edge cases.  The default code may work on other databases, such as MSSQL,
because it uses the SQL standard INFORMATION_SCHEMA.