Cache the database schema information by default, add the ability to get the schema for all tables at once
The method schema_for_table is renamed to schema, since it now can get
the schema for all tables in one call. To do so, you call the method
without arguments or with the table_name argument set to nil.  The
2nd argument is now an option hash.  The only option that all adapters
use is :reload, which always grabs fresh information from the database.
MySQL supports a :database option that chooses which database to use,
and PostgreSQL supports a :schema option that chooses which schema to
use.

Most of the private schema methods have been renamed and/or have had
been reworked.

The schema information is cached in the @schemas instance variable in
the database.