Refactor Inflections support, adding the inflection methods directly to String
Before, the inflection support was convoluted.  The String inflection
methods were in the Inflections module, which was included in String.
Each method in the Inflections module called an Inflector module
method, where Inflector was a module that extended itself.  Inside
Inflector was Inflector::Inflections, a class that included Singleton
and stored the singularization and pluralization rules.

This commit does away with a lot of that crap.  Inflector methods are
now String instance methods, and String::Inflections is a module
that operates as a singleton (similar to Sequel::PrettyTable).

Inflector = String, so if you use Inflector.inflections do..., it will
still work.  You should change that to String.inflections when you have
a chance.

Minor updates to the inflection documentation while I'm there.

Inflector#ordinalize didn't have an Inflections counterpart,
so it was deleted.