Refactor blockless filter support
Split ComplexExpression functionality into three subclasses:
BooleanExpression, NumericExpression, and StringExpression.  Each
subclass implements certain methods that the other subclasses do not.

BooleanExpression: &, |, ~
NumericExpression: +, -, *, /, <, >, <=, >=
StringExpression: like, <, >, <=, >=

The methods are defined in the BooleanMethods, NumericMethods,
StringMethods, and InequalityMethods modules, which are included
in the above classes as necessary.

All modules are included in the ComplexExpressionMethods module
which is included in LiteralString, Symbol, and SQL::Function.
ComplexExpressionMethods used to be included in SQL::Expression,
but it only really makes sense for SQL::Function, and would
have caused problems if it were included in SQL::Expression,
since ComplexExpression is a subclass of that.

ComplexExpressionMethods also has three new methods: sql_boolean,
sql_number, and sql_string, which return the receiver wrapped
in the appropriate type of expression.  This can be useful if
other libraries overwrite the operators for classes that Sequel
defines the SQL for.  For example, I ran into a situation where
Symbol#/ was defined to make file system path creation easier,
but it returned the result as a string, which Sequel produced
the wrong SQL with (quoting it as a string).  The fix was to
use :symbol.sql_number/10.