Have the PostgreSQL and MySQL adapters use the Sequel.time_class setting for datetime/timestamp types
Add a time type to the schema parser, that is used for database time
types (which have time but not date).  There isn't really a ruby
equivalent for this, as the time class holds the date as well, but
continue to use Time when returning and typecasting since it has
been used historical and there isn't a better class in ruby's
standard library.

Have PostgreSQL money type use BigDecimal instead of Float.

Support the PostgreSQL time with time zone type.