Add a :one_to_one option to one_to_many associations, which creates a getter and setter similar to many_to_one (a.k.a. has_one)
The usual use of this is to split one table into multiple tables,
generally either for security or performance reasons.

The getter returns a singular matching record, or raises an error if
multiple records match.

The setter updates the record given and removes associations with all
other records.

The usual many_to_one association methods are either removed if they
are not needed (remove_$, remove_all_$) or made private if they are
($, $_dataset, $_helper, add_$).

Assume that each song has only one lyric and each lyric has only one
song, but the table is split because you don't use the lyrics often
and it performs better to have them in a separate table:

  Song.one_to_many :lyrics, :one_to_one=>true # Note the plural
  song = Song.new(:id => 1)
  song.lyric # <Lyric...>, or nil if no association lyric
  lyric = Lyric.new(:id => 2)
  song.lyric = lyric # UPDATE lyrics SET song_id = 1 WHERE id = 2;
                     # UPDATE lyrics SET song_id = NULL WHERE id != 2
                     #   AND song_id = 1;

Also, this changes the code so that add_ and remove_ for one_to_many
associations call save instead of save! on the object, and raise an
error if the object couldn't be saved.  This was the intent when the
code was originally written.  save! skips validation, it doesn't
raise an error instead of returning false.