The expr argument in join table now allows the same argument as filter, so it can take a string or a blockless filter expression
The following use of join_table now means something different:

  DB[:artists].join_table(:inner, :albums, :artist_id)
  Before:
    => "INNER JOIN albums ON albums.artist_id = artist.id"
  After:
    => "INNER JOIN albums ON artist_id"

Obviously, that is a loss of functionality for that case, but this
commit allows the following cases:

  DB[:artists].join_table(:inner, :albums, "artist_id = artists.id")
  DB[:artists].join_table(:inner, :albums, :number_sold > 1000000)

That's right, you can now use blockless filters, strings, or anything
accepted by filter.

To fix your code that used the previous implementation:

  Change:
    DB[:artists].join_table(:inner, :albums, :artist_id)
  To:
    DB[:artists].join_table(:inner, :albums, :artist_id=>:id)

Coming soon will be the ability for join_table to take a block that
gives the table alias arguments, so you can create blockless filters
that work for arbitrary joins without knowing in advance what table
aliases will be used.