Support the SQL EXTRACT function: :date.extract(:year)
Stop abusing ColumnExpr to support the SQL CAST function. Just
because the SQL syntax is the same doesn't mention you should
use invalid sematics. This comes at a slight cost to backwards
compatibility according to the specs:

  Before:
    :x.cast_as(:varchar[20])
  Now:
    :x.cast_as('varchar(20)')

This is actually a good thing, as the before example is an abuse of
the SQL::Function class, when the result is not a function call but a
type specifier.

Add a SQL::IrregularFunction class to handle EXTRACT and CAST.

Add ComplexExpressionMethods#extract, for easily extracting parts
(e.g. month, minute) from dates/times.  Since these parts should
always be numeric, return the result as a NumericExpression.