Refactor SQL::Expression subclasses so that methods that don't make sense aren't defined
This refactor fixes some issues, like being able to do:

  :item.desc.as(:blah)

which makes no sense in SQL.

It adds a few new features.  One is a qualify method for symbols
that can add a table/schema qualifier for a column/table name:

  :price.qualify(:items) # items.price

Another are cast_numeric and cast_string, which will cast with
default types (integer and text) and convert the resulting value
to a NumericExpression or StringExpression, so operators that
don't make sense aren't defined (like &).  Example:

  :date.extract(:hour).cast_string + ':00'
  # CAST(extract(hour FROM date) AS text) || ':00'

Also, cast is now the preferred name for the method, instead of
cast_as (which is now an alias to cast).

The ColumnMethods module and ColumnExpr class are going away,
because they combine wildly different semantics with similar
syntax. They are being replaced with separate modules (AliasMethods,
OrderMethods, CastMethods) and classes (OrderedExpression,
AliasedExpression).

There should be very minimal fallout from this:

1) Inverting the order of a dataset will get you ASC if you had DESC,
instead of leaving ASC off.

2) Plain Strings no longer have asc and desc methods, since ordering
by a plain string is a noop.  LiteralStrings still have asc and desc
methods.

3) You can no longer abuse the SQL::Function DSL (Symbol#[]) to
use a table alias with specified columns.  I have no idea if anyone
used this, but it no longer works.  I doubt anyone needs it, so there
isn't a good replacement, other than LiteralString.

The new hierarchy looks like this:

Expression
    ComplexExpression
        BooleanExpression
        NumericExpression
        StringExpression
    GenericExpression
        CaseExpression
        Function
            IrregularFunction
        QualifiedIdentifier
        Subscript
    SpecificExpression
        AliasedExpression
        ColumnAll
        OrderedExpression