Store cached associated objects in @associations hash instead of one instance variable per association
This will make it so that you can use associations such as :values,
as well as other names of instance variables.  It also simplifies the
code somewhat, and exposes the cached associations hash via
attr_reader, so you can use it to easily check whether the associated
object(s) has been cached or not.

One small change for this is that the :reciprocal association object
now must provide the symbol name of the association instead of the
name of the instance variable (which was just the name prepended with
@).