Remove :pool_reuse_connections ConnectionPool/Database option, MySQL users need to be careful with nested queries
The :pool_reuse_connections option is being removed because it could
cause problems with transactions not getting the same database
connection.  The only reason it was added was to prevent MySQL users
from getting an error for this type of query:

  DB[:table1].each{|x| DB[:table2].each{|y| ... }}

MySQL leaves a database cursor open inside each, and it will raise a
"Commands out of sync" error for the above code.  The previous
workaround still works:

  DB[:table1].all{|x| DB[:table2].each{|y| ... }}

The fix is to use all instead of each for the surround query, so the
database query finishes executing before the first call to the block.