Rename Model.sti_key to Model.set_sti_key, and have it set up filters in subclasses
Model.set_sti_key now adds filters to subclasses.  This doesn't work
perfectly for class hierarchies with a depth greater than 2, though.
If you have subclasses of subclasses of the parent STI class, you'll
need to set up the dataset filter in the middle classes manually.

The filter that set_sti_key adds is a simple matching of the sti_key
to the class name.  The class that calls set_sti_key doesn't have
the filter applied to it, so it should get all subclasses.