Allow Dataset #filter, #or, #exclude and other methods that call them to use both the block and regular arguments
Before, the regular arguments would be silently ignored and only the
block arguments would be used.  Now, if both regular and block
arguments are given, both will be used (ANDed together).

Another change is that the following no longer raises an error:

  dataset.filter(true)
  dataset.filter(false)

Previously, this raised an error, the reason given was that it
appeared to be an error.  However, it's perfectly valid to use
a plain boolean condition in a filter call, especially false
as a way of making sure no records get returned (true is really
a no-op).