Add migrate options (-m, -M) to the sequel command line tool, and exit properly
The -m option specifies the directory of migration files.

The -M option specifies the version to migrate to.  There is no
option to specify the version to migrate from, you'll have to
use the API for that.

This also changes the exit status to 1 instead of the default 0 in
case of an error.