Use the safer connection-specific string escaping methods for PostgreSQL
Remove the Sequel::Postgres::Adapter.quote method, as well as the
TRUE, FALSE, and NULL constants in Sequel::Postgres::Adapter.  Handle
all of the escaping inside Sequel::Postgres::Dataset#literal.  There
shouldn't be any breakage from this, unless you were using the
Adapter.quote method directly.

It probably hurts performance slightly to use the connection-specific
string escape methods, since we have to acquire a connection for
every attempt to literalize a string, but safety is more important.

The one corner case I can think of where this approach may pose an
issue is if you cannot acquire a connection right away.  It will wait
for a connection to become available, or even throw a
PoolTimeoutError if it can't acquire a connection.  We might be able
to work around that by using any already allocated connection, but
I'm not sure if that is safe.