Make :many_to_one associations support :dataset, :order, :limit association options, as well as block arguments
This makes the many_to_one association code use the same
infrastructure as the other association code for getting records.
This also makes it create a _dataset method for many_to_one
associations, just as it does for other associations.  Because the
same infrastructure is used, it automatically supports the :order,
:limit (to specify offset), and :dataset options, as well as a block
argument.

This can be used to set up ActiveRecord has_one-style associations,
by using a :dataset option, instead of using a one_to_many
association with the :one_to_one option:

  Band.many_to_one :band_info, :key=>:id, \
    :dataset=>proc{BandInfo.filter(:band_id => id)}

This commit also refactors some private association methods to remove
the redundant name argument, since it is included in the
AssociationReflection.