Do not apply the :limit association option when eager loading
Using :limit when eager loading will lead to wrong results in most
cases, so ignore that option when eager loading.  Note that eager
loading on a limited dataset still works correctly, it is just the
:limit association option that is ignored.

Applying the :limit association option was recently added, but it
is obviously wrong. Not sure what I was thinking when I added it.