Add :before_add, :after_add, :before_remove, :after_remove association callback options
These mirror the association callback capabilities of ActiveRecord.
Each of these options can be a Symbol specifying an instance method
that takes one argument (the associated object), or a Proc that takes
two arguments (the current object and the associated object).

If any of the before callbacks return false, the adding/removing
does not happen and it either raises an error (the default), or
returns nil (if raise_on_save_failure is false).

All callbacks are also run on many_to_one associations.  If there
was already an existing object for the association, it calls the
remove callbacks on the existing object and the add callbacks on the
new object.  The remove callback calls are placed around the add
callback calls.

This also separates the many_to_one setter (association=) method into
two parts, the public association= method, and the private
_association= method that actually sets the foreign key.  This is
necessary to support polymorphic associations, where another column
would also need to be changed, and probably in other cases as well.

This also changes the many_to_one setter logic somewhat, as it
doesn't just look in the associations cache for an existing object,
it just calls the association method, which means using the setter
can cause database access.  This was necessary to make sure the
remove callback methods worked correctly for many_to_one
associations.

The many_to_one setter now exits early if the object passed to it
is the same as the existing object.