Make validation methods support a :tag option, to work correctly with source reloading
If a validation has a :tag option, it overwrites a previous
validation block with the same tag and attribute (if any), similar to
how tags work with hooks.

All validates_*_of methods set the tag to * by default (e.g.
validates_uniqueness_of uses a :tag of :uniqueness by default). This
breaks backwards compatibility slightly, if you have two of the same
type of validations on the same attribute.  If that is the case, add
a different tag to one or both.

While here, refactor the :if option handling to be less verbose.
Also, document all of the supported options to the validates_*_of
methods, as well as some other documentation improvements.