Remove ParseTree support, the expression syntax is now the only available syntax inside blocks
Sequel.use_parse_tree always returns false, and
Sequel.use_parse_tree= raises an Error if the argument given is true.

Other than the the obvious change to the syntax inside blocks, the
only significant change is that update no longer takes a block. As
update only makes sense for equality, a block would not add any value
over using a hash argument.  Before, it was necessary to take a block
to do things like x = x + 1, but that has been easily done with
{:x=>:x + 1} since Sequel 2.0.0.

No longer include the deprecation feature, but still leave it in the
repository for future use.