Cloned datasets no longer ignore the existing columns unless it is necessary
This bug took me a while to track down.  It shows how much Sequel
really needs an integration test suite (I actually found it while
working on a prototype for the test suite).  Because Dataset#columns
needs access to the database to get results, it is mocked in almost
all specs.

This bug didn't cause any problems, since it would always get the
same columns back from the database, but it did cause a lot of
pointless queries.