Only select table.* by default for many_to_many associations
For other association types, it is not needed, and using it caused
problems if you wanted to use a :dataset option that included
an eager_graph call.  Now select is only called on the dataset
if it is specifically set or many_to_many is used.  If you want
many_to_many to select all columns, you can use :select=>nil.

While here, add documentation for the callback options.