Add :after_load association callback
This lets you run code after the associated items have been loaded
from the database.

Also, this lets you specify a nil :key option for a many_to_one
association.  If specified the query will load all records for
the association, and just give you the first one.  This allows
many_to_one to work correctly with datasets that use eager_graph,
as well as with associations where there is no foreign key in the
current table, or when both are true (which is possible for a
has_one :through style association).