Make validates_length_of default tag depend on presence of options passed to it
This fixes the issue where prior calls to validates_length_of for the
same attribute get ignored.  That is true of the other validations as
well, but there is no reason to use any of them twice.
validates_length_of could easily get used twice, once for :maximum and
once for :minimum with different error messages specified for each,
for example.