Add integration test suite, testing sequel against a real database, with nothing mocked
After a long wait, I've finally put together an integration test
suite for sequel.  Currently, it only covers the examples shown in
the advanced_associations.rdoc file, but hopefully now that it is
here, more people will contribute real specs.

The integration specs have two ways of running, one where the
database is specified, and one where it is not.  If the database is
not specified, it uses a sqlite memory database, and addition to
checking for the correct functionality, it also checks that the
correct SQL is being issued to the server.  If the database is
specified, it uses that database and ignores the SQL (since it can
be different depending on the database).  This gives the best of
both worlds, the ability to check that sequel works with real
databases, and the ability to see what SQL sequel is actual issuing
to the server when nothing is mocked.  Incidently, a prototype of
this test suite is what helped me discover the Database#columns bug.

You can run the integration test suite with "rake integration", and
configure it via the INTEGRATION_URL constant in spec_config.rb, or
via the SEQUEL_INTEGRATION_URL environment variable.

While here, change the MYSQL_DB and POSTGRES_DB examples in
spec_config.rb to MYSQL_URL and POSTGRES_URL, as it is preferable to
use the URL instead of DB constant.