Typecast '' to nil by default for non-string non-blob columns, add typecast_empty_string_to_nil= model class and instance methods
This is mainly to ease web development using Sequel, where '' would
generally be the value returned by a select box for an integer column
when nothing is selected.  It will also help any time you are using
Sequel with values extracted out of a text based file (such as csv),
where the empty string for an integer field is used to express NULL.

I've given this quite a bit of thought about whether to make this
option true by default.  In general, Sequel does not go out of its
way to make web development easier, defaulting to options that are
strict and will generally be changed for web development, such as
strict_param_setting and typecast_on_assignment.  However, both of
those features are generally only changed if you want to display
nice error messages to the user, because they only change the
behavior when the submitted data is malformed.  In this case, Sequel
would be raising an error during normal use.  It's a tough call to
make, but ultimately I think that typecasting to nil is the better
default.

Like many of the other options, this is modifiable at the global,
class, and instance levels:

  Sequel::Model.typecast_empty_string_to_nil = false
  Artist.typecast_empty_string_to_nil = false
  artist = Artist.new
  artist.typecast_empty_string_to_nil = false