Add more integration tests, fix PostgreSQL insert without primary key
This adds a couple of integration tests for simple dataset operations
and for schema parsing, adding tables, and adding/removing columns
to/from tables.  I'm adding these now because there is going to be
some significant refactoring happening relatively soon and I want to
be extra sure that nothing gets broken.

The integration suite already found one bug with the PostgreSQL
adapter, that raised an error trying to insert a record into a table
without a primary key.  This commit fixes that.

This commit also makes the SQLite adapter not override the
alter_table add_column case, since the default does work.