Support JRuby 1.1.3, allow use of jdbc-postgres gem, fix some JDBC issues
This commit has a small restructuring of the JDBC adapter to allow it
to work with JRuby 1.1.3 (hopefully it still works on JRuby 1.1.2).
It makes sure the Database/Dataset is always correctly modified, by
extending it in Database#initialize instead of Database#connect. This
fixes issues such as creating a dataset before connecting to the
database not picking up the PostgreSQL specific settings.

This commit allows you to use the jdbc-postgres gem, but it still
works if you just have the PostgreSQL-JDBC .jar in your CLASSPATH.

This commit also fixes JDBC::Database#inspect, and handles the
literalization of Java::JavaSql::Timestamp values.