Enable full support for SQLite-JDBC using the JDBC adapter
SQLite-JDBC is now fully supported using the JDBC adapter.  It passes
the SQLite specs (which did have to be modified, but now support both
the native and JDBC adapters.  It passes the integration test suite.

This was done similar to the PostgreSQL-JDBC adapter, by spliting the
native adapter into shared and unshared parts, using the shared parts
in the JDBC adapter, and reimplementing the unshared parts.

This doesn't mean that the native and JDBC drivers operate exactly
the same.  For example, the native adapter will return strings for
values such as computed columns, where the JDBC adapter will return
numbers.  The JDBC adapter also does not allow submitting multiple
statements at once.  Also, the JDBC adapter will raise a generic
Error instead of Error::InvalidStatement if there is a problem with
the query.

One significant change is that Schema operations can now return
arrays in addition to strings, and the Database object will handle it
correctly.  This was done to allow drop_column to work with
SQLite-JDBC, since you can't submit multiple statements at once,
and that is necessary to handle a column drop in SQLite.

A type_test integration test has been added for testing that certain
database types are supported.

A few minor changes:

* JDBC::Database::url alias to uri was added
* Database::<< uses execute_ddl instead of execute
* The SQLite specs no longer require the use of a memory database
* You can use the SQLITE_URL constant instead of SQLITE_DB
* You can use the SEQUEL_SQLITE_SPEC_DB environment variable as well
* The dataset integration tests check delete and update return values