Enable almost full support for MySQL using JDBC
As much as it pained me, I did the necessary work splitting the MySQL
adapter into shared and native parts, and adding the necessary JDBC
support.  There are still three spec failures, one in the integration
specs and two in the MySQL specific specs.  One notable feature not
working is that MySQL via JDBC does not support the
convert_tinyint_to_boolean option.  I don't plan to improve the
MySQL JDBC situation further, but I'll accept patches to do so.

Now that the PostgreSQL, SQLite, and MySQL adapters have been split,
I don't plan to add support for other JDBC adapters.  It should be
fairly easy to add support for the remaining adapters using JDBC,
since all of the remaining adapters are fairly small.  If you get
another adapter working with JDBC, be sure to send a patch/pull
request.

There were a few related changes made to enable this that are worth
mentioning.  First is that JDBC::Database got a default transaction
method, which hopefully should work on most databases (it appears
to work with both SQLite and MySQL).  Second is that
Dataset#multi_insert now uses execute_dui instead of execute.