Huge changes, mostly to add prepared statement/bound variable support
This is a huge commit.  I general prefer to commit in smaller chunks,
but this is a major feature that will have a large effect on the
future of Sequel, and I didn't wan't to commit before I knew that
the code was flexible enough to work on multiple database types, and
that it didn't break existing code.

This commit adds support for prepared statements and bound variables.
Included is a prepared_statement.rdoc file, review that to get an
idea of usage.  This has been tested on PostgreSQL, MySQL, and
SQLite, both with the native drivers and with the JDBC drivers.  For
other databases, it emulates support using interpolation.

Along with the prepared statement support comes complete, but not
necessarily good, documentation for the PostgreSQL, MySQL, SQLite,
and JDBC adapters.

There were numerous minor changes made as well:

* MSSQL should be better supported on JDBC, though I haven't tested
  it.
* Statement execution on JDBC and SQLite was refactored to reduce
  code duplication.
* JDBC support for inserting records was refactored to reduce code
  duplication.
* Dataset got private execute and execute_dui methods, that send the
  the SQL to the database.  The adapters that had special database
  #execute methods had similar changes to their datasets.
* Mysql::Result#convert_type is now a private method.
* Mysql::Result#each_array was removed, probably a leftover from the
  old arrays with keys code.
* All databases now have a @transactions instance variable set on
  initialization, saving code inside #transaction.
* Native support for prepared statements when using PostgreSQL can be
  determined by seeing if SEQUEL_POSTGRES_USES_PG is defined and
  true.
* Postgres::Adapter#connected was removed.
* #serial_primary_key_options was removed from the MySQL and SQLite
  adapters, since it was the same as the default.
* Postgres::Database#locks was refactored.
* Use block_given? and yield instead of block[] in some places.
* Database#log_info takes an additional args argument, used for
  logging bound variables.
* The InvalidExpression, InvalidFilter, InvalidJoinType, and
  WorkerStop exceptions were removed.
* Using Model[true] and Model[false] no longer results in an error.
  This was probably more helpful with the ParseTree expression
  syntax (which used == for equality), but shouldn't be helpful now.
* Using a full outer join on MySQL raises a generic Error instead of
  InvalidJoinType.