In PostgreSQL adapter, fix inserting a row with a primary key value inside a transaction
This bug took quite a while to find and fix.  It was compounded by
the fact that there were multiple bugs in the underlying
implementation:

1) Errors raised by PostgreSQL weren't getting reraised inside
   insert_result.

2) Determining the sequence names was not done correctly.

This caused weird errors, such as the code raising an error stating
there was already an error on the transaction, without showing
an error or even that a query was issued.  To fix this and similar
issues, have the connection log the SQL it issues to find primary
keys, sequences, and sequence values, just like the Database object
logs the SQL.

To make sure that inserting a row with a primary key value works
inside transactions, figure out the primary key first and see if
it contained in the values hash.  If not, figure out what the
sequence is for the table, and get the last sequence value used.

This is done because trying to get the last sequence value first
if the sequence wasn't yet used on the connection causes PostgreSQL
to abort the transaction.  It would return an invalid result
instead of aborting the transaction if the connection had previously
been used to insert a row into the same table, without an easy way
to detect things.

Fix the SQL used for finding primary keys and sequences so that
unnecessary columns aren't returned.

Keep track of sequences at the Database level instead of the
connection level.

Make insert_result, primary_key_for_table, and
primary_key_sequence_for_table private methods.

The :table option when inserting is now unquoted (generally a symbol),
instead of the quoted string used previously.